Debug

+++ FF_HEDM/src/IndexerScanningOMP.c
@@ -1292,7 +1292,7 @@ int ReadSpots(char *cwd)
 	printf("%lld %lld\n",(long long int) size,(long long int)sizeof(double));
 	ObsSpotsLab = mmap(0,size,PROT_READ,MAP_SHARED,fd,0);
 	check (ObsSpotsLab == MAP_FAILED,"mmap %s failed: %s", filename, strerror(errno));
-	return (int) size/(10*sizeof(double));
+	return (int) size/((size_t)(10*sizeof(double)));
 	//~ return (int) size/(10*8);
 }
